Moved the comments domain from the posts package into apps/admin
ref https://linear.app/ghost/issue/PLA-221/posts-admin-move-tags-comments-automations

- third domain to leave the posts package as it dissolves into apps/admin; comments already read the filter engine from admin-x-framework (extracted earlier), so its only posts-internal ties were MainLayout and the comments-only pinning hook, both handled here
- the comments route folds into the shell's own route table; MainLayout is replaced with shade Box + Container, unchanged rendering
- use-comments-pinning-enabled moves into the comments domain (its only consumers)
- brings the migrated code up to Admin's stricter lint/tsconfig: verbatimModuleSyntax type-only imports, voided react-query fetchNextPage, a non-component thread-link helper extracted for react-refresh, and the nql-lang test type declaration carried over
- the moved tests colocate beside their sources; adds @tryghost/nql-lang to admin for the comment-fields filter test
