+++ apps/admin/package.json
@@ -19,6 +19,7 @@
         "@tryghost/admin-x-settings": "workspace:*",
         "@tryghost/kg-unsplash-selector": "workspace:*",
         "@tryghost/koenig-lexical": "workspace:*",
+        "@tryghost/nql-lang": "catalog:",
         "@tryghost/posts": "workspace:*",
         "@tryghost/shade": "workspace:*",
         "@xyflow/react": "12.11.1",
