+++ apps/admin/src/comments/components/comment-metrics.tsx
@@ -1,6 +1,7 @@
 import CommentLikesModal from './comment-likes-modal';
 import CommentReportsModal from './comment-reports-modal';
-import {Comment} from '@tryghost/admin-x-framework/api/comments';
+import {buildThreadLink} from './thread-link';
+import {type Comment} from '@tryghost/admin-x-framework/api/comments';
 import {Link, useSearchParams} from '@tryghost/admin-x-framework';
 import {LucideIcon, cn, formatNumber} from '@tryghost/shade/utils';
 import {Tooltip, TooltipContent, TooltipTrigger} from '@tryghost/shade/components';
@@ -62,18 +63,6 @@ function Metric({icon, count, label, to, onClick, className, testId}: MetricProp
             <TooltipContent>{isClickable ? `View ${label.toLowerCase()}` : label}</TooltipContent>
         </Tooltip>
     );
-}
-
-/**
- * Builds a thread link URL that preserves existing search params
- */
-export function buildThreadLink(searchParams: URLSearchParams, commentId: string | undefined | null): string | undefined {
-    if (!commentId) {
-        return undefined;
-    }
-    const newParams = new URLSearchParams(searchParams);
-    newParams.set('thread', `is:${commentId}`);
-    return `?${newParams.toString()}`;
 }
 
 interface CommentMetricsProps {
