+++ apps/admin/src/comments/components/comments-list.tsx
@@ -1,15 +1,16 @@
 import CommentContent from './comment-content';
 import CommentThreadSidebar from './comment-thread-sidebar';
 import {Avatar, Button, TooltipProvider} from '@tryghost/shade/components';
-import {Comment, useHideComment, useShowComment, useUnpinComment} from '@tryghost/admin-x-framework/api/comments';
+import {type Comment, useHideComment, useShowComment, useUnpinComment} from '@tryghost/admin-x-framework/api/comments';
 import {CommentHeader} from './comment-header';
 import {CommentMenu} from './comment-menu';
-import {CommentMetrics, buildThreadLink} from './comment-metrics';
+import {CommentMetrics} from './comment-metrics';
+import {buildThreadLink} from './thread-link';
 import {Link, useSearchParams} from '@tryghost/admin-x-framework';
 import {LoadMoreButton, useInfiniteVirtualScroll, useScrollRestoration, useVirtualListWindow} from '@tryghost/admin-x-framework/virtual-list';
 import {LucideIcon, cn} from '@tryghost/shade/utils';
 import {forwardRef, useEffect, useRef, useState} from 'react';
-import {useCommentsPinningEnabled} from '@src/hooks/use-comments-pinning-enabled';
+import {useCommentsPinningEnabled} from '@/comments/hooks/use-comments-pinning-enabled';
 
 const SpacerRow = ({height}: { height: number }) => (
     <div aria-hidden="true" className="flex">
