+++ apps/admin/src/comments/use-comment-filter-fields.ts
@@ -1,6 +1,6 @@
 import React, {useMemo} from 'react';
 import {DATE_OPERATOR_LABELS, RELATIVE_DATE_OPERATOR_LABELS, createOperatorOptions, createRelativeDateRenderer, fieldHasRelativeOperator, getTodayInTimezone} from '@tryghost/admin-x-framework/filters';
-import {FilterFieldConfig, ValueSource} from '@tryghost/shade/patterns';
+import type {FilterFieldConfig, ValueSource} from '@tryghost/shade/patterns';
 import {LucideIcon} from '@tryghost/shade/utils';
 import {commentFields} from './comment-fields';
