+++ apps/posts/src/routes.tsx
@@ -52,10 +52,6 @@ export const routes: RouteObject[] = [
                     }
                 ]
             },
-            {
-                path: 'comments',
-                lazy: lazyComponent(() => import('@views/comments/comments'))
-            },
 
             // Error handling
             {
